[probes] Remove common/statskeeper. (#801)
- Migrate the functionality to the only remaining user of this
  package - udplistener.
